Move Action class into ActionExecutor module
