+++ contextshell/CallableAction.py
@@ -1,6 +1,6 @@
 from typing import Callable
 from contextshell.backends.ActionExecutor import unpack_argument_tree, ActionArgsPack
-from contextshell.Action import Action
+from contextshell.backends.ActionExecutor import Action
 from contextshell.NodePath import NodePath
