+++ contextshell/backends/BuildinActionRegistryMixin.py
@@ -1,0 +1,3 @@
+
+class BuildinActionRegistryMixin:
+    pass
