+++ contextshell/backends/NodeTree.py
@@ -1,7 +1,7 @@
 from contextshell.Node import Node
 from contextshell.NodePath import NodePath
 from contextshell.backends.ActionEndpoint import ActionEndpoint
-from contextshell.Action import Action
+from contextshell.backends.ActionExecutor import Action
 from contextshell.CallableAction import action_from_function
 from typing import List, Optional
