+++ tests/unit/Fakes.py
@@ -1,5 +1,5 @@
 from contextshell.NodePath import NodePath
-from contextshell.Action import Action
+from contextshell.backends.ActionExecutor import Action
 from typing import Dict
 from contextshell.backends.ActionEndpoint import ActionArgsPack
